Separating the tests
